Improved DateFilter
Added an option to the DateFilter module that allows sorting of entries
by whether or not they are linked to another entry. This filter is
applied after the other filter, whatever it's setting was.

Also changed the "Linked" indicator, which was made redundant after the
GraphTools update which made the button text contextually sensitive, to
indicate how many entries are accessible after filters are applied.